refactor Gallery component to components folder
