+++ components/Gallery/index.js
@@ -1,0 +1,3 @@
+import Gallery from "./Gallery.js";
+
+export default Gallery;
